Added the following to inittab to fix an issue with lxc-stop on lxc 1.x and above. below is a note from a bug i had filed.
This is an issue with SIGPWR handling in the container.  The lxc-stop command now issues a SIGPWR signal to the container init and then waits a time out period for the container to gracefully shut down and exit.  The container init process must handle that signal and shut down gracefully.  If it fails to shut down gracefully before a timeout, lxc-stop will then kill the container.